mobile setup added on home view, database validations added, StyledBreadcrumb crrections, Breadcumb component deleted
