+++ client/src/Screens/AdminViews/Series/index.jsx
@@ -85,7 +85,6 @@ const EditSeries = ({
       },
     });
   };
-  console.log(series);
   return (
     <>
       <Filter
