+++ client/src/components/Slider/constants.js
@@ -1,0 +1,1 @@
+export const initialImage = "slide11";
