+++ client/src/translations/en-es/index.json
@@ -2,9 +2,9 @@
   "about.info": "Hola! {html} {html}  Soy un artista visual e ilustradora argentina dedicada al diseño y creación de piezas de arte únicas. {html} {html} Primero obtuve un título en Diseñado en comunicación visual. Posteriormente me convertí en artista plástico en la Universidad Nacional de La Plata. {html} En mis primeros años como profesional trabajé en diseño gráfico, abrí el camino a la producción artística y luego a la ilustración editorial. Domino las herramientas dentro de este campo y me encanta aplicarlas en busca de resultados innovadores. {html} He tenido la oportunidad de trabajar codo a codo con grandes artistas, lo que me permitió crecer profesionalmente. {html} {html} Me motiva la naturaleza, la poesía, la música, y la creatividad; me apasionan los colores, las líneas y las formas y anhelo lo sobrenatural o lo desconocido. {html} {html} Siempre he querido compartir mi magia. Sé que encontré cómo hacerlo.",
 
   "artwork.infoTitle": "Titulo:",
-  "artowork.printsAvailable": "COPIAS DISPONIBLES",
-  "artwork.originalOnSale": "ORIGINALES A LA VENTA",
-  "artwork.here": "AQUI",
+  "artwork.printsAvailable": "REPRODUCCION",
+  "artwork.originalOnSale": "ORIGINAL",
+  "artwork.here": " DISPONIBLE AQUI",
 
   "navbar.home": "Inicio",
   "navbar.about": "Sobre mi",
