+++ client/src/translations/en-us/index.json
@@ -2,9 +2,9 @@
   "about.info": " Hola! {html} {html} I am an argentinian visual artist and illustrator devoted to unique art piece design and creation. {html} {html} I firstly got a visual communication designer degree. After that I became a plastic artist at Universidad Nacional de La Plata college. {html} In my early years as a professional I worked on graphic design, I paved the way to artistic production and editorial illustration later. I master tools within this scope and I delight on applying them in search of innovative results. {html}I have had the chance to work with great artists side by side, which allowed me to professionally grow up every time. {html} {html} Being motivated by nature, poetry, music, words as well as creativity, I feel passionate about colors, lines and forms and long for the supernatural or the unknown. I always combine this all and apply it on new projects. {html} {html} I have always wanted to share my magic. I know I found how to do it.",
 
   "artwork.infoTitle": "Title:",
-  "artowork.printsAvailable": "PRINTS AVAILABLE",
-  "artwork.originalOnSale": "ORIGINAL ON SALE",
-  "artwork.here": "HERE",
+  "artwork.printsAvailable": "PRINT",
+  "artwork.originalOnSale": "ORIGINAL",
+  "artwork.here": "AVAILABLE HERE",
 
   "navbar.home": "Home",
   "navbar.about": "About",
